fix: audit-37 hardening (UART baud/divisor/port overflow, PCI MSI-X arbitrary-MMIO, loop/SB16) (#54)
* fix(drivers): audit-37 UART/PCI-MSI/loop/SB16 untrusted-input hardening

serial_8250.rs / serial_16550.rs / serial_console.rs / pl011_uart.rs:
- baud_divisor split the divide (UART_CLOCK_HZ/16/baud) so the intermediate
  16*baud cannot overflow u32 on an attacker baud >= 2^28 (HIGH ring-0 panic);
  serial_console.rs carried the same un-split multiply (verify residual).
- read_reg/write_reg compute base+offset in u32 before narrowing to the u16 I/O
  port so a base near 0xFFFF cannot wrap (MEDIUM).
- PL011 init/compute_divisors compute FBRD in u64 + clamp to 63, checked_mul for
  baud*16, and validate IBRD in [1,0xFFFF] (HIGH/MEDIUM overflow).

pci_msi.rs / pci_msi_irq.rs / pci_bridge.rs:
- MSI-X table accessors (msix_read/write_entry_field, write/read_msix_entry) take
  the real table size and reject an out-of-range entry before the MMIO pointer
  math = arbitrary-MMIO via an attacker vector index (CRITICAL/HIGH); callers
  thread the clamped num_vectors.
- MsixCapability::parse clamps the device-reported num_vectors to MAX_MSIX_VECTORS.
- bridge configure_windows disables a window when aligned_size==0 (CRITICAL
  unsigned underflow); align_up uses checked_add (HIGH); enumerate_bridge
  validates secondary>primary and subordinate>=secondary bus numbers.

loop_dev.rs / sb16_audio.rs / usb_serial.rs:
- loop read_block/write_block use checked_mul for lba*block_size (HIGH overflow);
  capacity_blocks guards block_size==0 (verify residual div0 on the pub field).
- SB16 init validates dma8 in {0,1,3} / dma16 in {5,6,7} before the 1<<dma8 shift
  and dma16-4 subtraction (HIGH shift/underflow).
- usb_serial set_config and set_baud_rate reject baud_rate==0 (LOW div0; the
  set_baud_rate doc promised a rejection the body did not perform — verify residual).

* fix(hal): audit-37 MSI-X table accessor index ceiling

msix_read_entry_field / msix_write_entry_field computed the table MMIO address
from a raw entry index whose only bound was an unsafe-fn doc contract. Add a
defense-in-depth ceiling (entry < MSIX_MAX_VECTORS = 2048, field within the
16-byte entry) so a contract violation driven by an attacker-influenced vector
index cannot synthesize an arbitrary out-of-table MMIO read/write primitive.
Surfaced by the audit-37 verify sibling sweep (the drivers-side copies were
fixed to take an explicit table size).

+++ crates/drivers/src/serial_16550.rs
@@ -584,12 +584,16 @@ impl Serial16550 {
 
     /// Read a UART register.
     fn read_reg(&self, offset: u16) -> u8 {
-        port_inb(self.base + offset)
+        // SECURITY: widen to u32 before adding so that a base near 0xFFFF
+        // cannot wrap the u16 sum; offsets are 0..=7 so the result is always
+        // a valid port for any real COM base address (<= 0x3F8 + 7).
+        port_inb((self.base as u32 + offset as u32) as u16)
     }
 
     /// Write a UART register.
     fn write_reg(&self, offset: u16, val: u8) {
-        port_outb(self.base + offset, val);
+        // SECURITY: same widening rationale as read_reg.
+        port_outb((self.base as u32 + offset as u32) as u16, val);
     }
 
     // -- Data transfer -----------------------------------------------------
